Closed #8 set temperature source as enum

+++ Almostengr.ThermometerPi.Api/Database/TemperatureRepository.cs
@@ -2,8 +2,8 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Almostengr.ThermometerPi.Api.Constants;
 using Almostengr.ThermometerPi.Api.DataTransferObject;
+using Almostengr.ThermometerPi.Api.Enums;
 using Almostengr.ThermometerPi.Api.Models;
 using Microsoft.EntityFrameworkCore;
 
@@ -31,7 +31,7 @@ namespace Almostengr.ThermometerPi.Api.Database
         public async Task<TemperatureDto> GetLatestExteriorReadingAsync()
         {
             return await _dbContext.TemperatureReadings
-                .Where(r => r.Source == TemperatureSource.Exterior && r.Timestamp >= DateTime.Now.AddHours(-2))
+                .Where(r => r.Source == (int) TemperatureSource.Exterior && r.Timestamp >= DateTime.Now.AddHours(-2))
                 .OrderByDescending(r => r.Timestamp)
                 .Select(t => t.AsDto())
                 .FirstOrDefaultAsync();
@@ -40,7 +40,7 @@ namespace Almostengr.ThermometerPi.Api.Database
         public async Task<TemperatureDto> GetLatestInteriorReadingAsync()
         {
             return await _dbContext.TemperatureReadings
-                .Where(r => r.Source == TemperatureSource.Interior && r.Timestamp >= DateTime.Now.AddHours(-2))
+                .Where(r => r.Source == (int) TemperatureSource.Interior && r.Timestamp >= DateTime.Now.AddHours(-2))
                 .OrderByDescending(r => r.Timestamp)
                 .Select(t => t.AsDto())
                 .FirstOrDefaultAsync();
@@ -67,7 +67,7 @@ namespace Almostengr.ThermometerPi.Api.Database
         public async Task<TemperatureDto> GetMinInteriorReadingAsync()
         {
             return await _dbContext.TemperatureReadings
-                .Where(r => r.Source == TemperatureSource.Interior)
+                .Where(r => r.Source == (int) TemperatureSource.Interior)
                 .OrderBy(r => r.TemperatureF)
                 .Select(t => t.AsDto())
                 .FirstOrDefaultAsync();
@@ -76,7 +76,7 @@ namespace Almostengr.ThermometerPi.Api.Database
         public async Task<TemperatureDto> GetMaxInteriorReadingAsync()
         {
             return await _dbContext.TemperatureReadings
-                .Where(r => r.Source == TemperatureSource.Interior)
+                .Where(r => r.Source == (int) TemperatureSource.Interior)
                 .OrderByDescending(r => r.TemperatureF)
                 .Select(t => t.AsDto())
                 .FirstOrDefaultAsync();
